Delete Contact

+++ bulk_mail/views.py
@@ -14,7 +14,7 @@ def index(request):
         selected_groups=[Group.objects.get(id=i) for i in selected_id]
         emails=((selected_group.emails.values_list('email_address', flat=True)) for selected_group in selected_groups)
         all_emails=[email for sublist in emails for email in sublist]
-        email_message = EmailMultiAlternatives(subject, message, to=all_emails)#Send Email to all emails 
+        email_message = EmailMultiAlternatives(subject, message, to=all_emails)
         email_message.content_subtype='html'
         email_message.send()
         message=Message.objects.create(
@@ -39,10 +39,14 @@ def create_user(request):
         form = UserForm()
     return render(request, 'create_contact.html', {'form': form})
 
-def delete_user(request, user_id):
-    user = User.objects.get(id = user_id)
-    user.delete()
-    return redirect(reverse("view_contacts"))
+def delete_user(request, id):
+    if request.GET:
+        id = request.GET.get('id')
+        user = User.objects.get(id = id)
+        user.delete()
+        return redirect(reverse("view_contacts"))
+    else:
+        HttpResponse("Operation Denied")
